backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
google-api-python-client
python-dotenv